% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_evaluate.R
\name{.classify_with_granges}
\alias{.classify_with_granges}
\title{Classify DMRs using GenomicRanges overlaps}
\usage{
.classify_with_granges(called_dmrs, truth_regions, min_overlap_bp)
}
\description{
Classify DMRs using GenomicRanges overlaps
}
\keyword{internal}
