% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_run_methods.R
\name{.merge_group_bedgraphs}
\alias{.merge_group_bedgraphs}
\title{Merge two group bedGraphs into metilene's expected format}
\usage{
.merge_group_bedgraphs(path_A, path_B, split_data, out_path)
}
\description{
Merge two group bedGraphs into metilene's expected format
}
\keyword{internal}
