% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_simulate_efficiency.R
\name{get_effect_sizes}
\alias{get_effect_sizes}
\title{Get predefined effect sizes for spike-in simulation}
\usage{
get_effect_sizes()
}
\value{
Numeric vector of effect sizes to test.
}
\description{
Get predefined effect sizes for spike-in simulation
}
