% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_additional_figures.R
\name{plot_call_breakdown}
\alias{plot_call_breakdown}
\title{Stacked bar chart showing composition of DMR calls (TP vs FP)}
\usage{
plot_call_breakdown(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table of spike-in results}

\item{out_dir}{Output directory}
}
\value{
ggplot object
}
\description{
Stacked bar chart showing composition of DMR calls (TP vs FP)
}
