% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_real_data_comparison.R
\name{plot_effect_size_distribution}
\alias{plot_effect_size_distribution}
\title{Effect size distribution per method}
\usage{
plot_effect_size_distribution(dmr_list, out_dir = NULL)
}
\description{
Effect size distribution per method
}
