% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_results.R
\name{plot_f1_heatmap}
\alias{plot_f1_heatmap}
\title{Heatmap of F1 scores across effect sizes and methods}
\usage{
plot_f1_heatmap(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table from compile_results().}

\item{out_dir}{Output directory.}
}
\value{
ggplot object
}
\description{
Heatmap of F1 scores across effect sizes and methods
}
