% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_additional_figures.R
\name{plot_f1_lines}
\alias{plot_f1_lines}
\title{F1 score as line plot, easier to read than the heatmap}
\usage{
plot_f1_lines(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table of spike-in results}

\item{out_dir}{Output directory}
}
\value{
ggplot object
}
\description{
F1 score as line plot, easier to read than the heatmap
}
