% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_results.R
\name{plot_fdr}
\alias{plot_fdr}
\title{FDR across methods and effect sizes}
\usage{
plot_fdr(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table from compile_results().}

\item{out_dir}{Output directory.}
}
\value{
ggplot object
}
\description{
FDR across methods and effect sizes
}
