% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_results.R
\name{plot_null_fp}
\alias{plot_null_fp}
\title{Bar plot of false positive DMR counts across scenarios and methods}
\usage{
plot_null_fp(null_dt, out_dir = NULL)
}
\arguments{
\item{null_dt}{data.table from compile_null_results()}

\item{out_dir}{Output directory for figure files.}
}
\value{
ggplot object
}
\description{
Bar plot of false positive DMR counts across scenarios and methods
}
