% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_results.R
\name{plot_sensitivity_curves}
\alias{plot_sensitivity_curves}
\title{Line plot of sensitivity vs effect size, faceted by efficiency scenario}
\usage{
plot_sensitivity_curves(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table from compile_results() for spike-in sim.}

\item{out_dir}{Output directory.}
}
\value{
ggplot object
}
\description{
Line plot of sensitivity vs effect size, faceted by efficiency scenario
}
