% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_additional_figures.R
\name{plot_sensitivity_precision_panels}
\alias{plot_sensitivity_precision_panels}
\title{Combined sensitivity and precision panels}
\usage{
plot_sensitivity_precision_panels(spikein_dt, out_dir = NULL)
}
\arguments{
\item{spikein_dt}{data.table of spike-in results}

\item{out_dir}{Output directory}
}
\value{
ggplot object
}
\description{
Combined sensitivity and precision panels
}
