% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_real_data_comparison.R
\name{plot_dmr_counts}
\alias{plot_dmr_counts}
\title{Bar plot of total DMR counts per method}
\usage{
plot_dmr_counts(dmr_list, out_dir = NULL)
}
\description{
Bar plot of total DMR counts per method
}
